Move datasets from inputters to datatypes.
