+++ onmt/datatypes/image_dataset.py
