+++ onmt/inputters/__init__.py
@@ -7,9 +7,9 @@ from onmt.inputters.inputter import make_features, \
     load_old_vocab, get_fields, OrderedIterator, \
     build_dataset, build_vocab, old_style_vocab
 from onmt.inputters.dataset_base import DatasetBase
-from onmt.inputters.text_dataset import TextDataset
-from onmt.inputters.image_dataset import ImageDataset
-from onmt.inputters.audio_dataset import AudioDataset
+from onmt.datatypes.text_dataset import TextDataset
+from onmt.datatypes.image_dataset import ImageDataset
+from onmt.datatypes.audio_dataset import AudioDataset
 
 
 __all__ = ['DatasetBase', 'make_features',
